refactor: extract content-block types to standalone search-domain types
Pulls ContentItem, DocumentReference, and DocumentMetadata out of
AgentInstanceHistoryEntity so they can also be reused by AgentInstanceEntity's
systemPrompt field.
