+++ search/search-domain/src/main/java/io/camunda/search/entities/DocumentReference.java
@@ -1,0 +1,24 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+package io.camunda.search.entities;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import java.util.Objects;
+import org.jspecify.annotations.Nullable;
+
+/** A document reference embedded in a {@link ContentItem}. */
+@JsonIgnoreProperties(ignoreUnknown = true)
+public record DocumentReference(
+    String storeId, String documentId, @Nullable String contentHash, DocumentMetadata metadata) {
+
+  public DocumentReference {
+    Objects.requireNonNull(storeId, "storeId");
+    Objects.requireNonNull(documentId, "documentId");
+    Objects.requireNonNull(metadata, "metadata");
+  }
+}
